CASE-1: Refactor Login test
