+++ TestRailAutomationTest/Page/BasePage.cs
@@ -1,4 +1,5 @@
 using OpenQA.Selenium;
+using TestRailAutomationTest.Utils;
 
 namespace TestRailAutomationTest.Page;
 
@@ -9,5 +10,18 @@ public abstract class BasePage
     protected BasePage(IWebDriver driver)
     {
         Driver = driver;
+    }
+    
+    public bool IsPageOpened(By uniqueElementLocation)
+    {
+        try
+        {
+            Waits.WaitElementExistence(Driver, uniqueElementLocation);
+            return true;
+        }
+        catch (NoSuchElementException)
+        {
+            return false;
+        }
     }
 }
