+++ TestRailAutomationTest/Resources/appsettings.json
@@ -1,6 +1,11 @@
 {
-  "HomePageUrl": "https://vadimevteev.testrail.io/",
-  "Browser": "chrome",
-  "Email": "vadzim.yautseyeu@itechart-group.com",
-  "Password": "gpNOOxLWntpkEDZ7KiPe"
+    "Browser": "chrome",
+    "AppUrl":   "https://vadimevteev.testrail.io/",
+    "DefaultTimeoutSeconds": 10,
+    "Users":[
+        {
+            "Email": "vadzim.yautseyeu@itechart-group.com",
+            "Password": "gpNOOxLWntpkEDZ7KiPe"
+        }
+    ]
 }
